requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.16.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
fastapi==0.103.2
greenlet==3.0.0
cryptography==41.0.3
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
mariadb==1.0.11
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==0.21.1
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.21
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2

typing_extensions==4.7.1
uuid==1.30
uvicorn==0.22.0